Obj Loader/Drawing + Camera Tweaks
Obj Loader:
Obj Loader is now working.
Separate draw method was added.
Sphere.obj loads fine.
PentaSphere.obj can only be loaded if Sphere.obj is loaded first.
This is odd since only one obj can be loaded at the moment and loading a
second overwrites the first.
Second oddity is that the mtl file must have a reference to a bmp file,
which was added for PentaSphere.
These files currently do nothing and while read in, the texture for
objects is not displayed as of this commit.

Camera:
'A' and 'D' move camera left and and right.
'W' and 'S' move camera up and down.
'Q' and 'R' move camera backwards and forwards.